+++ Runtime/NativeContainers/Array/NativeExchangeArray.cs
@@ -7,11 +7,13 @@ using Unity.Collections.LowLevel.Unsafe;
 namespace Voxell.NativeContainers
 {
   [StructLayout(LayoutKind.Sequential), NativeContainer, NativeContainerIsAtomicWriteOnly]
-  unsafe public struct NativeAdd : IDisposable
+  unsafe public struct NativeExchangeArray : IDisposable
   {
-    public bool IsCreated => _valueHolder != null;
+    public int Length => _length;
+    public bool IsCreated => _buffer != null;
 
-    [NativeDisableUnsafePtrRestriction] private int* _valueHolder;
+    [NativeDisableUnsafePtrRestriction] internal void* _buffer;
+    private int _length;
 
     #if ENABLE_UNITY_COLLECTIONS_CHECKS
     private AtomicSafetyHandle m_Safety;
@@ -20,55 +22,60 @@ namespace Voxell.NativeContainers
 
     private Allocator _allocatorLabel;
 
-    public NativeAdd(Allocator allocator)
+    public NativeExchangeArray(int length, Allocator allocator)
     {
-      // create a dispose sentinel to track memory leaks. This also creates the AtomicSafetyHandle
+      long size = UnsafeUtility.SizeOf<int>() * length;
+
       #if ENABLE_UNITY_COLLECTIONS_CHECKS
       if (allocator <= Allocator.None)
         throw new ArgumentException("Allocator must be Temp, TempJob or Persistent", nameof(allocator));
-      DisposeSentinel.Create(out m_Safety, out _disposeSentinel, 0, allocator);
+
+      if (length <= 0) throw new ArgumentOutOfRangeException(nameof(length), "Length must be > 0");
+      if (size > int.MaxValue)
+      {
+        throw new ArgumentOutOfRangeException(
+          nameof(length), $"Length * sizeof(int) cannot exceed {(object)int.MaxValue} bytes"
+        );
+      }
+      // create a dispose sentinel to track memory leaks. This also creates the AtomicSafetyHandle
+      DisposeSentinel.Create(out m_Safety, out _disposeSentinel, 1, allocator);
       #endif
 
-      // allocate native memory for a single integer
-      _valueHolder = (int*)UnsafeUtility.Malloc(UnsafeUtility.SizeOf<int>(), UnsafeUtility.AlignOf<int>(), allocator);
+      // allocate native memory for an array of integers
+      _buffer = UnsafeUtility.Malloc(size, UnsafeUtility.AlignOf<int>(), allocator);
+      _length = length;
       _allocatorLabel = allocator;
-
-      // initialize the value to 0 to avoid uninitialized data
-      Value = 0;
     }
 
-    public int Value
+    public int this[int index]
     {
       get
       {
-        // verify that the caller has read permission on this data
         #if ENABLE_UNITY_COLLECTIONS_CHECKS
         AtomicSafetyHandle.CheckReadAndThrow(m_Safety);
         #endif
-        return *_valueHolder;
+        return UnsafeUtility.ReadArrayElement<int>(_buffer, index);
       }
 
       [WriteAccessRequired]
       set
       {
-        // verify that the caller has write permission on this data
         #if ENABLE_UNITY_COLLECTIONS_CHECKS
         AtomicSafetyHandle.CheckWriteAndThrow(m_Safety);
         #endif
-        *_valueHolder = value;
+        UnsafeUtility.WriteArrayElement<int>(_buffer, index, value);
       }
     }
 
     [WriteAccessRequired]
-    public void AtomicAdd(int value)
+    public int AtomicExchange(int index, int value)
     {
       // verify that the caller has write permission on this data
       #if ENABLE_UNITY_COLLECTIONS_CHECKS
       AtomicSafetyHandle.CheckWriteAndThrow(m_Safety);
       #endif
-      Interlocked.Add(ref *_valueHolder, value);
+      return Interlocked.Exchange(ref *((int*)_buffer + index), value);
     }
-
 
     public void Dispose()
     {
@@ -77,8 +84,9 @@ namespace Voxell.NativeContainers
       DisposeSentinel.Dispose(ref m_Safety, ref _disposeSentinel);
       #endif
 
-      UnsafeUtility.Free(_valueHolder, _allocatorLabel);
-      _valueHolder = null;
+      UnsafeUtility.Free(_buffer, _allocatorLabel);
+      _buffer = null;
+      _length = 0;
     }
   }
 }