+++ Runtime/NativeContainers/NativeExchange.cs
@@ -9,32 +9,23 @@ namespace Voxell.NativeContainers
   [StructLayout(LayoutKind.Sequential), NativeContainer, NativeContainerIsAtomicWriteOnly]
   unsafe public struct NativeExchange : IDisposable
   {
-    // the actual pointer to the allocated count needs to have restrictions relaxed
-    // so jobs can be schedled with this container
+    public bool IsCreated => _valueHolder != null;
+
     [NativeDisableUnsafePtrRestriction] private int* _valueHolder;
 
     #if ENABLE_UNITY_COLLECTIONS_CHECKS
     private AtomicSafetyHandle m_Safety;
-    /*
-    the dispose sentinel tracks memory leaks, it is a managed type so it is cleared to null when scheduling a job
-    the job cannot dispose the container, and no one else can dispose it until the job has run,
-    so it is ok to not pass it along
-    this attribute is required, without it this NativeContainer cannot be passed to a job
-    since that would give the job access to a managed object
-    */
-    [NativeSetClassTypeToNullOnSchedule]
-    private DisposeSentinel _disposeSentinel;
+    [NativeSetClassTypeToNullOnSchedule] private DisposeSentinel _disposeSentinel;
     #endif
 
-    // Keep track of where the memory for this was allocated
     private Allocator _allocatorLabel;
 
     public NativeExchange(Allocator allocator)
     {
-      // create a dispose sentinel to track memory leaks. This also creates the AtomicSafetyHandle
       #if ENABLE_UNITY_COLLECTIONS_CHECKS
       if (allocator <= Allocator.None)
         throw new ArgumentException("Allocator must be Temp, TempJob or Persistent", nameof(allocator));
+      // create a dispose sentinel to track memory leaks. This also creates the AtomicSafetyHandle
       DisposeSentinel.Create(out m_Safety, out _disposeSentinel, 0, allocator);
       #endif
 
@@ -44,15 +35,6 @@ namespace Voxell.NativeContainers
 
       // initialize the value to 0 to avoid uninitialized data
       Value = 0;
-    }
-
-    public int AtomicExchange(int value)
-    {
-      // verify that the caller has write permission on this data
-      #if ENABLE_UNITY_COLLECTIONS_CHECKS
-      AtomicSafetyHandle.CheckWriteAndThrow(m_Safety);
-      #endif
-      return Interlocked.Exchange(ref *_valueHolder, value);
     }
 
     public int Value
@@ -65,6 +47,8 @@ namespace Voxell.NativeContainers
         #endif
         return *_valueHolder;
       }
+
+      [WriteAccessRequired]
       set
       {
         // verify that the caller has write permission on this data
@@ -75,7 +59,15 @@ namespace Voxell.NativeContainers
       }
     }
 
-    public bool IsCreated => _valueHolder != null;
+    [WriteAccessRequired]
+    public int AtomicExchange(int value)
+    {
+      // verify that the caller has write permission on this data
+      #if ENABLE_UNITY_COLLECTIONS_CHECKS
+      AtomicSafetyHandle.CheckWriteAndThrow(m_Safety);
+      #endif
+      return Interlocked.Exchange(ref *_valueHolder, value);
+    }
 
     public void Dispose()
     {